requirements.txt
mysqlclient
Django
djangorestframework
djangorestframework-simplejwt
django-model-utils